+++ packages/data-source/src/lib/index.ts
@@ -1,2 +1,0 @@
-export { FetchController } from './classes';
-export * from './types';