+++ packages/garden/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.2](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-garden@0.7.1...@equinor/workspace-garden@0.7.2) (2023-01-13)
+
+**Note:** Version bump only for package @equinor/workspace-garden
+
 ## [0.7.1](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-garden@0.7.0...@equinor/workspace-garden@0.7.1) (2022-12-12)
 
 **Note:** Version bump only for package @equinor/workspace-garden